client-application/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "cython>=3.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "licensed-app"
version = "1.0.0"
description = "Licensed Application with Cython Protection"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.20.0",
]

[tool.setuptools.packages.find]
include = ["controllers*", "models*", "services*", "utils*", "core*"]
